+++ delay.h
@@ -1,7 +1,9 @@
 
+#include <stdio.h>
+
 #ifndef DELAY_H_
 #define DELAY_H_
 
-void delay(unsigned int time);
+void delay(size_t time);
 
 #endif
